v1.0_beta1

+++ app/src/main/java/com/example/user/photocollecting/adapter/LibraryListViewAdapter.java
@@ -2,6 +2,7 @@ package com.example.user.photocollecting.adapter;
 
 import android.content.Context;
 import android.content.Intent;
+import android.graphics.Bitmap;
 import android.os.Environment;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -11,11 +12,13 @@ import android.widget.BaseAdapter;
 import android.widget.ImageView;
 import android.widget.LinearLayout;
 import android.widget.ListView;
+import android.widget.ProgressBar;
 import android.widget.RelativeLayout;
 import android.widget.TextView;
 import android.widget.Toast;
 
 import com.example.user.photocollecting.R;
+import com.example.user.photocollecting.Util.Constants;
 import com.example.user.photocollecting.entity.Goods;
 import com.example.user.photocollecting.view.DetailActivity;
 
@@ -42,12 +45,19 @@ public class LibraryListViewAdapter extends BaseAdapter{
         mListView.setOnItemClickListener(new AdapterView.OnItemClickListener() {
             @Override
             public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
-                Intent intent = new Intent(mContext, DetailActivity.class);
-                intent.putExtra("path", mGoodsList.get(position).getImgFile().getAbsolutePath());
-                intent.putExtra("name", mGoodsList.get(position).getName());
-                mContext.startActivity(intent);
+                if(mGoodsList.get(position).getImgFile() != null ){
+                    Intent intent = new Intent(mContext, DetailActivity.class);
+                    intent.putExtra("path", mGoodsList.get(position).getImgFile().getAbsolutePath());
+                    intent.putExtra("name", mGoodsList.get(position).getName());
+                    mContext.startActivity(intent);
+                }
+
             }
         });
+    }
+
+    public void setmGoodsList(List<Goods> mGoodsList) {
+        this.mGoodsList = mGoodsList;
     }
 
     @Override
@@ -75,18 +85,35 @@ public class LibraryListViewAdapter extends BaseAdapter{
             holder.img = (ImageView) convertView.findViewById(R.id.img);
             holder.name = (TextView) convertView.findViewById(R.id.name);
             holder.lay_swipe_delete = (RelativeLayout)convertView.findViewById(R.id.lay_swipe_delete);
-
+            holder.mProgressBar = (ProgressBar)convertView.findViewById(R.id.progressBar);
+            holder.mProgressBar.setMax(Constants.Num_Frame);
+            holder.lay_load = (LinearLayout)convertView.findViewById(R.id.lay_load);
             convertView.setTag(holder);//绑定ViewHolder对象
         } else{
             holder = (ViewHolder) convertView.getTag();//取出ViewHolder对象                  }
         }
 
-        holder.img.setImageBitmap(mGoodsList.get(position).getBitmap());
+        Bitmap bitmap = mGoodsList.get(position).getBitmap();
+        if(bitmap != null){
+            holder.img.setImageBitmap(bitmap);
+        }else {
+            holder.img.setImageResource(R.mipmap.img_default);
+        }
         holder.name.setText(mGoodsList.get(position).getName());
+        if(mGoodsList.get(position).isProcessing()){
+            holder.lay_load.setVisibility(View.VISIBLE);
+            holder.mProgressBar.setProgress(mGoodsList.get(position).getProgress());
+            holder.name.setVisibility(View.GONE);
+        }else {
+            holder.lay_load.setVisibility(View.GONE);
+            holder.name.setVisibility(View.VISIBLE);
+        }
+
         holder.lay_swipe_delete.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                deleteFile(mGoodsList.get(position).getName(), position);
+                if(!mGoodsList.get(position).isProcessing())
+                    deleteFile(mGoodsList.get(position).getName(), position);
             }
         });
         return convertView;
@@ -118,5 +145,7 @@ public class LibraryListViewAdapter extends BaseAdapter{
         ImageView img;
         TextView name;
         RelativeLayout lay_swipe_delete;
+        ProgressBar mProgressBar;
+        LinearLayout lay_load;
     }
 }
